second custom script used in Tilakaratna & Bensasson
Takes the output of chrbychr.pl as its input, and parses the summary of
bootstrap support to identify clades with at least 70% bootstrap
support. Tests whether a strain shares a clade with reference strains
from only a single population or subpopulation. Outputs Supplemental
File 2, and some of the data for Table 2 of Tilakaratna and Bensasson.